requirements.txt
kafka-python-ng
pynput
flask
requests
cryptography
python-dotenv